Minor change to unit test for class DataSet. DataSet test passed. This fixes #54 Changed unit test for class Event to use global program parameters. Event test passed. This fixes #53

+++ src/unitTests/DataSet_test.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-size_t DataSetTestSize = 100000;
+uint32_t DataSetTestSize = 100000;
 
 class DataSetTest : public ::testing::Test
 {
